+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 This log documents significant changes for each release.  This project follows
 [Semantic Versioning](http://semver.org/).
 
+## [3.13.4] - 2024-06-13
+### Fixed
+- a bug that could cause the context input parameter containing environment
+  variables to change.
+
 ## [3.13.3] - 2024-05-24
 ### Changed
 - Added separate TypeScript type definition files for the main file and each